pyproject.toml
[tool.poetry]
name = "personal-app"
version = "0.0.0"
description = "Simple set of applications, which are meant to simplify my life"
authors = ["Pavel Paranin <pavel@paran.in>"]
readme = "README.md"
packages = [{include = "personal_app"}]

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "^0.20.0"
pytest = "^7.2.0"
psycopg2 = "^2.9.5"
orjson = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
